requirements.txt
streamlit==1.32.0
tensorflow-cpu==2.9.1
h5py>=3.1.0,<3.10.0
numpy>=1.19.5,<1.25.0
Pillow>=9.0.0,<10.0.0
psutil==5.9.5
pandas>=1.3.0,<2.0.0
matplotlib>=3.5.0,<3.8.0
scikit-learn>=1.0.0,<1.3.0
protobuf>=3.20.0,<4.0.0
tqdm>=4.64.0
requests>=2.27.0